Extract nockma evaluator to a separate file
