+++ src/Juvix/Compiler/Nockma/Evaluator/Error.hs
@@ -1,0 +1,19 @@
+module Juvix.Compiler.Nockma.Evaluator.Error where
+
+import Juvix.Prelude hiding (Atom)
+import Juvix.Prelude.Pretty
+
+data NockEvalError
+  = InvalidPosition
+  | InvalidEncodedPosition
+  | InvalidOpCode
+  | ExpectedAtom
+  | ExpectedCell
+  deriving stock (Show)
+
+newtype GenericNockEvalError = GenericNockEvalError
+  { _genericNockEvalErrorMessage :: AnsiText
+  }
+
+class ToGenericNockEvalError a where
+  toGenericNockEvalError :: a -> GenericNockEvalError
